backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Google Gemini AI (2026 - google.genai SDK)
google-genai>=1.0.0

# Image processing
Pillow>=10.2.0

# Data validation
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Code quality and formatting
ruff>=0.4.0
mypy>=1.8.0
